+++ R/Plots.R
@@ -1,19 +1,22 @@
-
-#' Plot wrapper intended for use in a SpaDES module
+#' \code{Plot} wrapper intended for use in a SpaDES module
 #'
 #' This is a single function call that allows a module to change which format in which
-#' the plots will occur. Specifically, the two primary formats would be to "screen" or
-#' to disk as an image file, such as "png". THIS CURRENTLY ONLY WORKS CORRECTLY WITH
-#' ggplot2 objects that can be saved. It uses \code{Plot} internally, so individual
-#' plots may be rearranged. This function requires at least 2 things: a plotting function
-#' and data for that plotting function. See below and examples. THIS IS STILL EXPERIMENTAL and could change in the next release.
+#' the plots will occur.
+#' Specifically, the two primary formats would be to "screen" or to disk as an image file,
+#' such as "png".
+#' \emph{THIS CURRENTLY ONLY WORKS CORRECTLY WITH \code{ggplot2} objects that can be saved.}
+#' It uses \code{Plot} internally, so individual plots may be rearranged.
+#' This function requires at least 2 things: a plotting function and data for that plot function.
+#' See below and examples.
+#'
+#' @note THIS IS STILL EXPERIMENTAL and could change in the next release.
 #'
 #' This offers up to 4 different actions for a given plot:
 #'     \itemize{
 #'       \item To screen device
 #'       \item To disk as raw data
 #'       \item To disk as a saved plot object
-#'       \item To disk as a png or other "image" file, like pdf
+#'       \item To disk as a \file{.png} or other image file, e.g., \file{.pdf}
 #'     }
 #' To turn off plotting both to screen and disk, set both
 #' \code{.plotInititalTime = NA} and \code{.plots = NA} or any other
@@ -52,10 +55,10 @@
 #'   \item \code{type}
 #'     \itemize{
 #'       \item \code{"screen"} -- Will plot to the current device, normally a plot window
-#'       \item \code{"object"} -- Will save the plot object, e.g., ggplot object
+#'       \item \code{"object"} -- Will save the plot object, e.g., \code{ggplot} object
 #'       \item \code{"raw"} -- Will save the raw data prior to plotting, e.g.,
 #'                           the data argument
-#'       \item \code{"png"} -- or any other type saveble with \code{ggsave}
+#'       \item \code{"png"} -- or any other type save-able with \code{ggsave}
 #'     }
 #' }
 #'