+++ R/Plots.R
@@ -11,6 +11,11 @@
 #'
 #' @note THIS IS STILL EXPERIMENTAL and could change in the next release.
 #'
+#' `Plots` now has experimental support for "just a `Plot` call", ut with `types` specified.
+#' See example.
+#' The devices to save on disk will have some different behaviours to the screen representation,
+#' since "wiping" an individual plot on a device doesn't exist for a file device.
+#'
 #' This offers up to 4 different actions for a given plot:
 #'     \itemize{
 #'       \item To screen device
@@ -40,6 +45,10 @@
 #'   compatibility. A developer should set in the module to the intended initial
 #'   plot time and leave it.
 #' @param ggsaveArgs An optional list of arguments passed to \code{ggplot2::ggsave}
+#' @param deviceArgs An optional list of arguments passed to one of \code{png},
+#'       \code{pdf}, \code{tiff}, \code{bmp}, or \code{jgeg}. This is useful when
+#'       the plotting function is not creating a ggplot object.
+#'
 #' @param usePlot Logical. If \code{TRUE}, the default, then the plot will occur
 #'   with `quickPlot::Plot`, so it will be arranged with previously existing plots.
 #'
@@ -90,12 +99,17 @@
 #'       .plotInitialTime = 1
 #'       )
 #'
-#'  } # end of dontrun
+#' # Can also be used like quickPlot::Plot, but with control over output type
+#' r <- raster::raster(extent(0,10,0,10), vals = sample(1:3, size = 100, replace = TRUE))
+#' Plots(r, types = c("screen", "png"), deviceArgs = list(width = 700, height = 500))
+#'
+#' } # end of dontrun
 Plots <- function(data, fn, filename,
                   types = quote(params(sim)[[currentModule(sim)]]$.plots),
                   path = quote(file.path(outputPath(sim), "figures")),
                   .plotInitialTime = quote(params(sim)[[currentModule(sim)]]$.plotInitialTime),
                   ggsaveArgs = list(), usePlot = TRUE,
+                  deviceArgs = list(),
                   ...) {
 
   if (any(is(types, "call") || is(path, "call") || is(.plotInitialTime, "call"))){
@@ -105,6 +119,16 @@ Plots <- function(data, fn, filename,
       if (is(simIsIn, "try-error"))
         simIsIn <- NULL
     }
+  }
+
+  # Deal with non sim cases
+  if (is.null(simIsIn)) {
+    if (is.call(types) && any(grepl("sim", types)))
+      types <- "screen"
+    if (is.call(path) && any(grepl("sim", path)))
+      path = "."
+    if (is.call(.plotInitialTime) && any(grepl("sim", .plotInitialTime)))
+      .plotInitialTime <- 0L
   }
 
   if (!is.null(simIsIn))
@@ -125,28 +149,44 @@ Plots <- function(data, fn, filename,
     .plotInitialTime <- 0L
   }
 
-  ggplotClassesCanHandle <- c("eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg", "wmf")
   ggplotClassesCanHandleBar <- paste(ggplotClassesCanHandle, collapse = "|")
   needSave <- any(grepl(paste(ggplotClassesCanHandleBar, "|object"), types))
   needScreen <- !is.na(.plotInitialTime) && any(grepl("screen", types))
-  if (needScreen || needSave) {
-    gg <- fn(data, ...)
-    ggListToScreen <- setNames(list(gg), "gg")
-    if (!is.null(gg$labels$title) && needScreen) {
-      ggListToScreen <- setNames(ggListToScreen, gg$labels$title)
-      ggListToScreen[[1]]$labels$title <- NULL
+  if (missing(fn) && isTRUE(usePlot)) {
+    fn <- Plot
+  }
+  fnIsPlot <- identical(fn, Plot)
+  if (fnIsPlot) {
+    # make dummies
+    gg <- 1
+    ggListToScreen <- list()
+  } else {
+    if ( (needScreen || needSave) ) {
+      gg <- fn(data, ...)
+      if (!is(gg, ".quickPlot")) {
+        ggListToScreen <- setNames(list(gg), "gg")
+        if (!is.null(gg$labels$title) && needScreen) {
+          ggListToScreen <- setNames(ggListToScreen, gg$labels$title)
+          ggListToScreen[[1]]$labels$title <- NULL
+        }
+      }
     }
   }
 
   if (needScreen) {
-    if (is(gg, "gg"))
-      if (!requireNamespace("ggplot2")) stop("Please install ggplot2")
-    if (usePlot) {
-      names(ggListToScreen) <- gsub(names(ggListToScreen), pattern = " ", replacement = "_")
-      Plot(ggListToScreen, addTo = gg$labels$title)
+    if (fnIsPlot) {
+      gg <- fn(data, ...)
     } else {
-      print(gg)
+      if (is(gg, "gg"))
+        if (!requireNamespace("ggplot2")) stop("Please install ggplot2")
+      if (usePlot) {
+        names(ggListToScreen) <- gsub(names(ggListToScreen), pattern = " ", replacement = "_")
+        Plot(ggListToScreen, addTo = gg$labels$title)
+      } else {
+        print(gg)
+      }
     }
+
   }
   needSaveRaw <- any(grepl("raw", types))
   if (needSave || needSaveRaw) {
@@ -178,15 +218,38 @@ Plots <- function(data, fn, filename,
   }
 
   if (needSave) {
-    ggSaveFormats <- intersect(ggplotClassesCanHandle, types)
-    for (ggsf in ggSaveFormats) {
-      if (!requireNamespace("ggplot2")) stop("To save gg objects, need ggplot2 installed")
-      args <- list(plot = gg,
-                   filename = file.path(path, paste0(filename, ".", ggsf)))
-      if (length(ggsaveArgs)) {
-        args <- modifyList(args, ggsaveArgs)
+    if (is.null(simIsIn)) {
+      if (is.call(path))
+        path <- "."
+      if (is.call(path))
+        path <- "."
+    }
+    if (fnIsPlot) {
+      baseSaveFormats <- intersect(baseClassesCanHandle, types)
+      for (bsf in baseSaveFormats) {
+        type <- get(bsf)
+        theFilename <- file.path(path, paste0(filename, ".", bsf))
+        do.call(type, modifyList(list(theFilename), deviceArgs))
+        # curDev <- dev.cur()
+        clearPlot()
+        plotted <- try(fn(data, ...)) # if this fails, catch so it can be dev.off'd
+        dev.off()
+        if (!is(plotted, "try-error"))
+          message("Saved figure to: ", theFilename)
       }
-      do.call(ggplot2::ggsave, args = args)
+    } else {
+      ggSaveFormats <- intersect(ggplotClassesCanHandle, types)
+      for (ggsf in ggSaveFormats) {
+        theFilename <- file.path(path, paste0(filename, ".", ggsf))
+        if (!requireNamespace("ggplot2")) stop("To save gg objects, need ggplot2 installed")
+        args <- list(plot = gg,
+                     filename = theFilename)
+        if (length(ggsaveArgs)) {
+          args <- modifyList(args, ggsaveArgs)
+        }
+        do.call(ggplot2::ggsave, args = args)
+        message("Saved figure to: ", theFilename)
+      }
     }
 
     if (any(grepl("object", types)))
@@ -215,3 +278,4 @@ anyPlotting <- function(.plots) {
 }
 
 ggplotClassesCanHandle <- c("eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg", "wmf")
+baseClassesCanHandle <- c("pdf", "jpeg", "png", "tiff", "bmp")